moving form update logic to entity form TODO improve comparing of values
