+++ src/app/core/entity-components/entity-details/form/form.component.html
@@ -37,4 +37,5 @@
   [columns]="columns"
   [columnHeaders]="headers"
   [form]="form"
+  [saveInProgress]="saveInProgress"
 ></app-entity-form>
